// front/src/types/api.ts
export type IRegUser = {
  surname: string;
  middlename: string;
  username: string;
  email: string;
  birthday?: number;
  password: string;
  age?: boolean;
  gender: string;
  currency?: string; // Optional - backend automatically uses LU currency
  currency_temp?: any;
};